Cache negotiated api microversion for this server
If we negotiate an API microversion in the ironiccclient, cache
this value so that we don't need to re-negotiate it again for the
next API call, if the user doesn't specify a version.

We cache the version for a particular ironic instance in a
temporary file, one per server, in a multi-user, multi-server
safe-way - deleting old files once they expire, so only the
latest is kept.

Change-Id: I0232bc611789fb96491d855020042e23ba0c4fab
Blueprint: api-microversions